// ClientApp/src/app/shared/models/user.ts
export class User {
  id: number;
  no: number;
  name: string;
  surname: string;
  gender: number;
  dateOfBirth: Date;
  createdDate: Date;
  lastActiveDate: Date;
  introduction: string;
  fullAddress: string;
  hobbies: string;
  imageUrl: string;
  userName: string;
  email: string;
  phoneNumber: string;
  city: string;
  country: string;
  age: number;
  streetName: string;
  zipCode: string;
  genderName: string;
}